First batch of changes to permit VS2017 ARM64 tools to build required elements.

Related work items: #14268301

+++ ArmPkg/Drivers/ArmGic/ArmGicLib.c
@@ -134,8 +134,8 @@ ArmGicSendSgiTo (
 {
   MmioWrite32 (
     GicDistributorBase + ARM_GIC_ICDSGIR,
-    ((TargetListFilter & 0x3) << 24) | ((CPUTargetList & 0xFF) << 16) | SgiId
-    );
+	(UINT32)(((TargetListFilter & 0x3) << 24) | ((CPUTargetList & 0xFF) << 16) | SgiId)
+	);  // MS_CHANGE
 }
 
 /*
@@ -260,7 +260,7 @@ ArmGicEnableInterrupt (
   UINTN                 GicCpuRedistributorBase;
 
   // Calculate enable register offset and bit position
-  RegOffset = Source / 32;
+  RegOffset = (UINT32)(Source / 32);    // MS_CHANGE
   RegShift = Source % 32;
 
   Revision = ArmGicGetSupportedArchRevision ();
@@ -304,7 +304,7 @@ ArmGicDisableInterrupt (
   UINTN                 GicCpuRedistributorBase;
 
   // Calculate enable register offset and bit position
-  RegOffset = Source / 32;
+  RegOffset = (UINT32)(Source / 32);   // MS_CHANGE
   RegShift = Source % 32;
 
   Revision = ArmGicGetSupportedArchRevision ();
@@ -348,7 +348,7 @@ ArmGicIsInterruptEnabled (
   UINT32                Interrupts;
 
   // Calculate enable register offset and bit position
-  RegOffset = Source / 32;
+  RegOffset = (UINT32)(Source / 32);  // MS_CHANGE
   RegShift = Source % 32;
 
   Revision = ArmGicGetSupportedArchRevision ();
